+++ wdl-lint/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+* Added Trailing Comments lint rule as an extension of the `Whitespace` rule ([]())
+
 ## 0.6.0 - 09-16-2024
 
 ### Fixed